refactor templating code, add new subgraph for chainlink specific data that we track, update the README
