+++ mustache/templates/chainlink/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  description: 'Chainlink Rates API',
+  schemaFile: './chainlink.graphql',
+};
