--- mustache/templates/exchanger/set-start-blocks.js
@@ -1,78 +1,0 @@
-module.exports = {
-  Exchanger: {
-    prod: 10557958,
-    test: null,
-  },
-  Exchanger_v2: {
-    prod: 10772592,
-    test: null,
-  },
-  Exchanger_v3: {
-    prod: 11012438,
-    test: null,
-  },
-  Exchanger_v4: {
-    prod: 11239644,
-    test: null,
-  },
-  Exchanger_v5: {
-    prod: 11247235,
-    test: null,
-  },
-  Exchanger_v6: {
-    prod: 11363405,
-    test: null,
-  },
-  Synthetix: {
-    prod: 10782000,
-    test: null,
-  },
-  RatesBlockOne: {
-    prod: 10537958,
-    test: null,
-  },
-  RatesBlockTwo: {
-    prod: 10773070,
-    test: null,
-  },
-  RatesBlockThree: {
-    prod: 10873070,
-    test: null,
-  },
-  RatesBlockFour: {
-    prod: 10921231,
-    test: null,
-  },
-  RatesBlockFive: {
-    prod: 10950006,
-    test: null,
-  },
-  RatesBlockSix: {
-    prod: 10960006,
-    test: null,
-  },
-  RatesOIL_3: {
-    prod: 11198828,
-    test: null,
-  },
-  RatesAAVE_3: {
-    prod: 11179792,
-    test: null,
-  },
-  RatesDOT_3: {
-    prod: 11322483,
-    test: null,
-  },
-  RatesUNI_3: {
-    prod: 11322507,
-    test: null,
-  },
-  RatesFIL_3: {
-    prod: 11322375,
-    test: null,
-  },
-  RatesYFI_3: {
-    prod: 11322433,
-    test: null,
-  },
-};
