+++ package.json
@@ -10,9 +10,11 @@
   },
   "scripts": {
     "lint": "eslint .",
+    "yaml:chainlink": "node ./create-yaml-file create-yaml --subgraph chainlink --env prod --universal-test-block null | mustache - mustache/yaml.mustache > subgraphs/synthetix-chainlink.yaml",
     "yaml:rates": "node ./create-yaml-file create-yaml --subgraph rates --env prod --universal-test-block null | mustache - mustache/yaml.mustache > subgraphs/synthetix-rates.yaml",
     "yaml:exchanger": "node ./create-yaml-file create-yaml --subgraph exchanger --env prod --universal-test-block null | mustache - mustache/yaml.mustache > subgraphs/synthetix-exchanger.yaml",
     "codegen": "node prepare-abis && graph codegen",
+    "codegen:chainlink": "npm run yaml:chainlink && npm run codegen -- subgraphs/synthetix-chainlink.yaml",
     "codegen:snx": "npm run codegen -- subgraphs/synthetix.yaml",
     "codegen:rates": "npm run yaml:rates && npm run codegen -- subgraphs/synthetix-rates.yaml",
     "codegen:depot": "npm run codegen -- subgraphs/synthetix-depot.yaml",
@@ -25,6 +27,7 @@
     "codegen:liquidations": "npm run codegen -- subgraphs/synthetix-liquidations.yaml",
     "build": "graph build",
     "build:snx": "graph build subgraphs/synthetix.yaml",
+    "build:chainlink": "graph build subgraphs/synthetix-chainlink.yaml",
     "build:rates": "graph build subgraphs/synthetix-rates.yaml",
     "build:depot": "graph build subgraphs/synthetix-depot.yaml",
     "build:exchanges": "graph build subgraphs/synthetix-exchanges.yaml",
@@ -36,6 +39,7 @@
     "build:liquidations": "graph build subgraphs/synthetix-liquidations.yaml",
     "deploy": "graph deploy --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs/ --access-token $THEGRAPH_SNX_ACCESS_TOKEN",
     "deploy:snx": "npm run deploy -- synthetixio-team/synthetix subgraphs/synthetix.yaml",
+    "deploy:chainlink": "npm run deploy -- synthetixio-team/synthetix-chainlink subgraphs/synthetix-chainlink.yaml",
     "deploy:rates": "npm run deploy -- synthetixio-team/synthetix-rates subgraphs/synthetix-rates.yaml",
     "deploy:depot": "npm run deploy -- synthetixio-team/synthetix-depot subgraphs/synthetix-depot.yaml",
     "deploy:exchanges": "npm run deploy -- synthetixio-team/synthetix-exchanges subgraphs/synthetix-exchanges.yaml",
