+++ subgraphs/synthetix-chainlink.graphql
@@ -1,0 +1,19 @@
+# A chainlink Answer
+type AggregatorAnswer @entity {
+  id: ID!
+  currencyKey: Bytes!
+  synth: String!
+  aggregator: Bytes!
+  rate: BigInt!
+  roundId: BigInt!
+  block: BigInt!
+  timestamp: BigInt!
+}
+
+type LatestRate @entity {
+  " Name of synth. E.g. sUSD "
+  id: ID!
+
+  " Synth USD rate with 18 decimals "
+  rate: BigInt!
+}
